+++ stdlib/tests/crypto/mod.rs
@@ -7,7 +7,6 @@ mod ecdsa_secp256k1;
 mod elgamal;
 mod fri;
 mod keccak256;
-mod keccak256_deprecated;
 mod rpo;
 mod sha256;
 mod stark;